Added a request factory for the Android side of the lib.
