+++ android/src/main/kotlin/com/bruno/native_admob_flutter/rewarded_interstitial/Controller.kt
@@ -1,7 +1,7 @@
 package com.bruno.native_admob_flutter.rewarded_interstitial
 
 import android.app.Activity
-import com.bruno.native_admob_flutter.NativeAdmobFlutterPlugin
+import com.bruno.native_admob_flutter.RequestFactory
 import com.bruno.native_admob_flutter.encodeError
 import com.google.android.gms.ads.AdError
 import com.google.android.gms.ads.FullScreenContentCallback
@@ -30,8 +30,8 @@ class RewardedInterstitialController(
             "loadAd" -> {
                 channel.invokeMethod("loading", null)
                 val unitId: String = call.argument<String>("unitId")!!
-                val nonPersonalizedAds = call.argument<Boolean>("nonPersonalizedAds")
-                RewardedInterstitialAd.load(context, unitId, NativeAdmobFlutterPlugin.createAdRequest(nonPersonalizedAds), object : RewardedInterstitialAdLoadCallback() {
+                val nonPersonalizedAds = call.argument<Boolean>("nonPersonalizedAds")!!
+                RewardedInterstitialAd.load(context, unitId, RequestFactory.createAdRequest(nonPersonalizedAds), object : RewardedInterstitialAdLoadCallback() {
                     override fun onAdLoaded(ad: RewardedInterstitialAd) {
                         rewardedInterstitialAd = ad
                         channel.invokeMethod("onAdLoaded", null)
